- Move controller code to ProjectController from UserController - Fix mapping - Add resource result return to project like endpoints for testing purposes.
